Move pick close/reopen actions to groups/[id]/categories/[categoryId] route
Relocates the category detail page (with pick close/reopen functionality)
from the standalone /categories/[id] route to the architecturally correct
/groups/[id]/categories/[categoryId] path. The pick detail page
(HeartButton, OptionList) is moved to the corresponding
groups/[id]/categories/[categoryId]/picks/[pickId] route.

Closes #126
