--- src/app/categories/[id]/copy.ts
@@ -1,11 +1,0 @@
-export const CATEGORY_DETAIL_COPY = {
-  closePickButton: "Close Pick",
-  closePickError: "Unable to close this pick.",
-  closedBadge: "Closed",
-  noPicksMessage: "No picks have been added to this category yet.",
-  picksLabel: "Picks",
-  reopenPickButton: "Reopen",
-  errors: {
-    reopenFailed: "Failed to reopen pick. Please try again.",
-  },
-} as const;
