--- src/app/categories/[id]/picks/[pickId]/copy.ts
@@ -1,17 +1,0 @@
-export const PICK_DETAIL_COPY = {
-  addOptionButton: "Add",
-  addOptionPlaceholder: "Add an option...",
-  adoptButton: "Use this",
-  closedNotice: "This Pick is closed. Preferences can no longer be changed.",
-  errors: {
-    default: "Something went wrong. Please try again.",
-  },
-  heart: {
-    addOwnership: "Add to your options",
-    closed: "This Pick is closed",
-    removeOwnership: "Remove from your options",
-  },
-  noOptionsMessage: "No options have been added yet.",
-  optionsHeading: "Options",
-  suggestionsHeading: "Suggestions from your prior picks",
-} as const;
