+++ src/app/groups/[id]/categories/[categoryId]/copy.ts
@@ -1,4 +1,13 @@
 export const CATEGORY_DETAIL_COPY = {
+  closePickButton: "Close Pick",
+  closePickError: "Unable to close this pick.",
+  closedBadge: "Closed",
   createdAtLabel: "Created",
+  errors: {
+    reopenFailed: "Failed to reopen pick. Please try again.",
+  },
+  noPicksMessage: "No picks have been added to this category yet.",
   notFound: "Category not found.",
+  picksLabel: "Picks",
+  reopenPickButton: "Reopen",
 } as const;
