+++ src/app/groups/[id]/categories/[categoryId]/page.tsx
@@ -1,10 +1,13 @@
+import { revalidatePath } from "next/cache";
 import { notFound, redirect } from "next/navigation";
 
 import { getCategoryById } from "@/server/data/categories";
 import { getGroupById } from "@/server/data/groups";
+import { closePick, getPicksByCategory } from "@/server/data/picks";
 import { getVerifiedUid } from "@/server/utils/auth";
 
 import { CategoryDetailView } from "./CategoryDetailView";
+import { CATEGORY_DETAIL_COPY } from "./copy";
 
 export default async function CategoryDetailPage({
   params,
@@ -19,10 +22,45 @@ export default async function CategoryDetailPage({
 
   if (!group?.memberIds.includes(uid)) notFound();
 
-  const category = await getCategoryById(categoryId);
+  const [category, picks] = await Promise.all([
+    getCategoryById(categoryId),
+    getPicksByCategory(categoryId),
+  ]);
 
   if (!category) notFound();
   if (category.groupId !== id) notFound();
 
-  return <CategoryDetailView category={category} />;
+  async function closePickAction(formData: FormData) {
+    "use server";
+
+    const actionUid = await getVerifiedUid();
+    if (!actionUid) redirect("/sign-in");
+
+    const latestGroup = await getGroupById(id);
+    if (!latestGroup?.memberIds.includes(actionUid)) {
+      throw new Error(CATEGORY_DETAIL_COPY.closePickError);
+    }
+
+    const latestCategory = await getCategoryById(categoryId);
+    if (latestCategory?.groupId !== id) {
+      throw new Error(CATEGORY_DETAIL_COPY.closePickError);
+    }
+
+    const pickId = formData.get("pickId");
+    if (typeof pickId !== "string" || !pickId.trim()) {
+      throw new Error(CATEGORY_DETAIL_COPY.closePickError);
+    }
+
+    await closePick(categoryId, pickId);
+
+    revalidatePath(`/groups/${id}/categories/${categoryId}`);
+  }
+
+  return (
+    <CategoryDetailView
+      category={category}
+      closePickAction={closePickAction}
+      picks={picks}
+    />
+  );
 }
