+++ src/app/groups/[id]/categories/[categoryId]/picks/[pickId]/OptionList.spec.tsx
