+++ src/app/groups/[id]/categories/[categoryId]/picks/[pickId]/OptionListView.tsx
