+++ src/app/groups/[id]/categories/[categoryId]/picks/[pickId]/copy.ts
@@ -1,3 +1,21 @@
+export const PICK_DETAIL_COPY = {
+  addOptionButton: "Add",
+  addOptionPlaceholder: "Add an option...",
+  adoptButton: "Use this",
+  closedNotice: "This Pick is closed. Preferences can no longer be changed.",
+  errors: {
+    default: "Something went wrong. Please try again.",
+  },
+  heart: {
+    addOwnership: "Add to your options",
+    closed: "This Pick is closed",
+    removeOwnership: "Remove from your options",
+  },
+  noOptionsMessage: "No options have been added yet.",
+  optionsHeading: "Options",
+  suggestionsHeading: "Suggestions from your prior picks",
+} as const;
+
 export const PICK_DETAIL_SCAFFOLD_COPY = {
   closedStatusChip: "Closed",
   openStatusChip: "Open",
